Move Module's ImportResolver to imports.
