+++ src/interpreter/module.rs
@@ -3,8 +3,7 @@ use std::cell::RefCell;
 use std::fmt;
 use std::collections::HashMap;
 use std::borrow::Cow;
-use elements::{External, FunctionType, GlobalType, InitExpr, Internal, MemoryType, Opcode,
-			   ResizableLimits, TableType, Type};
+use elements::{External, FunctionType, InitExpr, Internal, Opcode, ResizableLimits, Type};
 use interpreter::{Error, MemoryInstance, RuntimeValue, TableInstance};
 use interpreter::imports::{ImportResolver, Imports};
 use interpreter::global::{GlobalInstance, GlobalRef};
@@ -525,68 +524,6 @@ impl<'a> NotStartedModuleRef<'a> {
 	pub fn assert_no_start(self) -> Result<ModuleRef, Error> {
 		assert!(self.validated_module.module().start_section().is_none());
 		Ok(self.instance)
-	}
-}
-
-impl ImportResolver for ModuleInstance {
-	fn resolve_func(
-		&self,
-		field_name: &str,
-		_func_type: &FunctionType,
-	) -> Result<FuncRef, Error> {
-		Ok(self.export_by_name(field_name)
-			.ok_or_else(|| {
-				Error::Validation(format!("Export {} not found", field_name))
-			})?
-			.as_func()
-			.ok_or_else(|| {
-				Error::Validation(format!("Export {} is not a function", field_name))
-			})?)
-	}
-
-	fn resolve_global(
-		&self,
-		field_name: &str,
-		_global_type: &GlobalType,
-	) -> Result<GlobalRef, Error> {
-		Ok(self.export_by_name(field_name)
-			.ok_or_else(|| {
-				Error::Instantiation(format!("Export {} not found", field_name))
-			})?
-			.as_global()
-			.ok_or_else(|| {
-				Error::Instantiation(format!("Export {} is not a global", field_name))
-			})?)
-	}
-
-	fn resolve_memory(
-		&self,
-		field_name: &str,
-		_memory_type: &MemoryType,
-	) -> Result<MemoryRef, Error> {
-		Ok(self.export_by_name(field_name)
-			.ok_or_else(|| {
-				Error::Instantiation(format!("Export {} not found", field_name))
-			})?
-			.as_memory()
-			.ok_or_else(|| {
-				Error::Instantiation(format!("Export {} is not a memory", field_name))
-			})?)
-	}
-
-	fn resolve_table(
-		&self,
-		field_name: &str,
-		_table_type: &TableType,
-	) -> Result<TableRef, Error> {
-		Ok(self.export_by_name(field_name)
-			.ok_or_else(|| {
-				Error::Instantiation(format!("Export {} not found", field_name))
-			})?
-			.as_table()
-			.ok_or_else(|| {
-				Error::Instantiation(format!("Export {} is not a table", field_name))
-			})?)
 	}
 }
